README: Actually list the pubkey fingerprint
The README specified what the fingerprint of the public key should be
but actually listed the key instead of its fingerprint.

I copied the key and generated the fingerprint using the following
command. This should be easily reproducible.

```
> ssh-keygen -l -f darwin-build-box.pub
256 SHA256:9YrbQYYrKN/2GKwFgJz0+fpSbaGe1lJuYFQyPMVdwCY no comment (ED25519)
```

I left the actual public key in the README because there's no harm in
doing so but in principle it might not be necessary.